feat: Custom status code and location header
- status code can be set in configuration
- response contains Location header

+++ src/config.rs
@@ -15,7 +15,7 @@ pub struct Config {
 pub struct ConfigRedirectRecord {
     pub re: String,
     pub rewrite_rule: String,
-    pub status_code: u8,
+    pub status_code: u16,
 }
 
 impl Config {
